property keys extracted to a separate class
