+++ src/main/java/com/studerw/tda/client/TdaClientException.java
@@ -1,0 +1,16 @@
+package com.studerw.tda.client;
+
+public class TdaClientException extends RuntimeException {
+
+  public TdaClientException(String message) {
+    super(message);
+  }
+
+  public TdaClientException(Throwable cause) {
+    super(cause);
+  }
+
+  public TdaClientException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
