+++ src/main/java/com/studerw/tda/client/TdaClientProperty.java
@@ -1,0 +1,13 @@
+package com.studerw.tda.client;
+
+public final class TdaClientProperty {
+
+  public static final String ACCESS_TOKEN = "tda.token.access";
+  public static final String REFRESH_TOKEN = "tda.token.refresh";
+  public static final String CLIENT_ID = "tda.client_id";
+  public static final String TDA_URL = "tda.url";
+  public static final String DEBUG_BYTES_LENGTH = "tda.debug.bytes.length";
+
+  private TdaClientProperty() {
+  }
+}
