--- src/main/java/com/studerw/tda/constants/TdaConstants.java
@@ -1,9 +1,0 @@
-package com.studerw.tda.constants;
-
-public interface TdaConstants {
-
-    String ACCESS_TOKEN = "tda.token.access";
-    String REFRESH_TOKEN = "tda.token.refresh";
-    String CLIENT_ID = "tda.client_id";
-    String TDA_URL = "tda.url";
-}
